refactor: `test_mora_to_text.py` 移植
