+++ test/tts_pipeline/test_text_analyzer.py
@@ -7,6 +7,7 @@ from voicevox_engine.tts_pipeline.text_analyzer import (
     Label,
     MoraLabel,
     UtteranceLabel,
+    mora_to_text,
     text_to_accent_phrases,
 )
 
@@ -318,6 +319,31 @@ class TestUtteranceLabel(TestBaseLabels):
     def test_features(self):
         """UtteranceLabel に含まれる features をテスト"""
         self.assertEqual(features(self.utterance_hello_hiho), self.test_case_hello_hiho)
+
+
+class TestMoraToText(TestCase):
+    def test_voice(self):
+        self.assertEqual(mora_to_text("a"), "ア")
+        self.assertEqual(mora_to_text("i"), "イ")
+        self.assertEqual(mora_to_text("ka"), "カ")
+        self.assertEqual(mora_to_text("N"), "ン")
+        self.assertEqual(mora_to_text("cl"), "ッ")
+        self.assertEqual(mora_to_text("gye"), "ギェ")
+        self.assertEqual(mora_to_text("ye"), "イェ")
+        self.assertEqual(mora_to_text("wo"), "ウォ")
+
+    def test_unvoice(self):
+        self.assertEqual(mora_to_text("A"), "ア")
+        self.assertEqual(mora_to_text("I"), "イ")
+        self.assertEqual(mora_to_text("kA"), "カ")
+        self.assertEqual(mora_to_text("gyE"), "ギェ")
+        self.assertEqual(mora_to_text("yE"), "イェ")
+        self.assertEqual(mora_to_text("wO"), "ウォ")
+
+    def test_invalid_mora(self):
+        """変なモーラが来ても例外を投げない"""
+        self.assertEqual(mora_to_text("x"), "x")
+        self.assertEqual(mora_to_text(""), "")
 
 
 def _gen_mora(text: str, consonant: str | None, vowel: str) -> Mora:
